Move navigation scripts into their own directory & namespace
